+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cobaltcore-dev/dashboard
 
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies [00bfb76]
+  - @cobaltcore-dev/aurora@0.9.0
+
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [bae772e]
+  - @cobaltcore-dev/aurora@0.8.1
+
 ## 1.1.3
 
 ### Patch Changes